removed extra files
