--- Trial 2.0/navbar.html
@@ -1,9 +1,0 @@
-<div class="topnav">
-    <img src="images\BillLogoUpdated.png" />
-    <a href='index.html'>Home</a>
-    <a href='VsBillTitlePage.html'>Vs. Bill</a>
-    <a class="active" href='sologamescreen.html'>Solo</a>
-
-    <a href="howtoplay.html">How to Play</a>
-  <input id="authorizeBtn" type="button" value=" Login "><br />
-</div>
