--- Trial 2.0/solointro.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <link rel="stylesheet" href="outline.css" >
-    <link rel="stylesheet" href="gameintropage.css">
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Welcome to Solo Mode!</title>
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Nunito&display=swap" rel="stylesheet">
-  </head>
-  <body>
-    <div class="topnav">
-        <a href="index.html"> <img src="images\BillLogoUpdated.png" /> </a>
-        <a href='index.html'>Home</a>
-        <a href='VsBillTitlePage.html'>Vs. Bill</a>
-        <a class="active" href='solointro.html'>Solo</a>
-        <a href="howtoplay.html">How to Play</a>
-      <input id="authorizeBtn" type="button" value=" Login "><br />
-    </div>
-
-    <div class="gametitle">Solo Mode</div>
-
-    <div class="description">
-      How well do you know your own music taste? 
-      Guess your own favorite songs, albums, and genres in single player mode!
-    </div>
-    <div class="play-wave">
-      <img class="wave" src="images/leftsoundwavesolidorange.png"/>
-      <a href = "sologamescreen.html" style="text-decoration: none;">
-        <button class = "playbutton">Play!</button>
-      </a>
-      <img class="wave" src="images/rightsoundwavesolidorange.png"/>
-    </div>
-    
-    
-  </body>
-</html>
