--- Trial 2.0/vsbillgamescreen.css
@@ -1,37 +1,0 @@
-.gametitle {
-    font-size: 13rem;
-    text-align: center;
-    margin-top: 1rem;
-    color: white;
-}
-
-input[type=text]{
-    width: 60%;
-    margin: auto;
-    box-sizing: border-box;
-    background-color: #FFC7AF;
-    color: black;
-    height: 5rem;
-    display:flex;
-    align-items: center;
-    justify-content: center;
-    margin-top: 10rem;
-    font-size: 2rem;
-    border-radius: 25%;
-    border-color: black;
-}
-
-input[type=submit] {
-    width: 20%;
-    margin: auto;
-    box-sizing: border-box;
-    align-items: center;
-    justify-content: center;
-    display: flex;
-    height: 4rem;
-    margin-top: 5rem;
-    background-color: #1D8C9F;
-    color: black;
-    border-radius: 25%;
-    font-size: 2rem;
-}
