--- Trial 2.0/vsbillgamescreen.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <link rel="stylesheet" href="outline.css" >
-    <link rel="stylesheet" href="vsbillgamescreen.css">
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Welcome to vs. Bill!</title>
-  </head>
-  <body>
-     <div class="topnav">
-        <img src="images\BillLogoUpdated.png" />
-        <a href='index.html'>Home</a>
-        <a class="active" href='VsBillTitlePage.html'>Vs. Bill</a>
-        <a href='sologamescreen.html'>Vs. Yourself</a>
-        <a href="howtoplay.html">How to Play</a>
-     </div>
-
-     <div class="gametitle">Question #1</div>
-
-     <form onsubmit="return addAnswers">
-        <input type="text" id="q1">
-        <input type="submit" value = "Submit">
-     </form>
-  </body>
-</html>
