--- Trial 2.0/vsbillgamescreen.js
@@ -1,9 +1,0 @@
-//size 10 array with answers from game mode
-const billanswers [];
-
-function addAnswers() {
-    var ans1 = document.getElementById('q1').value 
-    billanswers.push(ans1);
-    console.log(billanswers);
-    return false;
-}
