+++ public/authorization_button.js
