+++ public/images/blueSoundWavesLeft.png
