+++ public/images/blueSoundWavesRight.png
