+++ public/images/hollowOrangeSoundWaveLeft.png
