+++ public/navbar.js
