+++ public/sologamescreen.css
