+++ public/sologamescreen.html
