+++ public/sologamescreen.js
